Authentication required for all CRUDs

+++ app/controllers/GroupController.php
@@ -9,10 +9,8 @@ class GroupController extends BaseController {
 
     public function index()
     {
-        //if (Auth::check()) {
-        if (true) {
+        if (Auth::check()) {
             // The user is logged in...
-
             $groups = Group::all();
             return View::make('groups.index')
                 ->with('groups', $groups)
@@ -33,8 +31,14 @@ class GroupController extends BaseController {
     public function create()
     {
         // load the create form (app/views/groups/create.blade.php)
-        return View::make('groups.create')
-            ->with('title', 'Group create');
+        if (Auth::check()) {
+            return View::make('groups.create')
+                ->with('title', 'Group create');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -44,27 +48,33 @@ class GroupController extends BaseController {
      */
     public function store()
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'name'      => 'required',
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'name'      => 'required',
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('groups/create')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('groups/create')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $group = new Group;
+                $group->name       = Input::get('name');
+                $group->save();
+
+                // redirect
+                Session::flash('message', 'Successfully created group');
+                return Redirect::to('groups');
+            }
         } else {
-            // store
-            $group = new Group;
-            $group->name       = Input::get('name');
-            $group->save();
-
-            // redirect
-            Session::flash('message', 'Successfully created group');
-            return Redirect::to('groups');
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
         }
     }
 
@@ -76,13 +86,19 @@ class GroupController extends BaseController {
      */
     public function show($id)
     {
-        // get the group
-        $group = Group::find($id);
+        if (Auth::check()) {
+            // get the group
+            $group = Group::find($id);
 
-        // show the view and pass the group to it
-        return View::make('groups.show')
-            ->with('group', $group)
-            ->with('title', 'Group show');
+            // show the view and pass the group to it
+            return View::make('groups.show')
+                ->with('group', $group)
+                ->with('title', 'Group show');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -93,13 +109,19 @@ class GroupController extends BaseController {
      */
     public function edit($id)
     {
-        // get the group
-        $group = Group::find($id);
+        if (Auth::check()) {
+            // get the group
+            $group = Group::find($id);
 
-        // show the edit form and pass the group
-        return View::make('groups.edit')
-            ->with('group', $group)
-            ->with('title', 'Group edit');
+            // show the edit form and pass the group
+            return View::make('groups.edit')
+                ->with('group', $group)
+                ->with('title', 'Group edit');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -110,28 +132,35 @@ class GroupController extends BaseController {
      */
     public function update($id)
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'name'       => 'required'
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'name'       => 'required'
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('groups/' . $id . '/edit')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('groups/' . $id . '/edit')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $group = Group::find($id);
+                $group->name       = Input::get('name');
+                $group->save();
+
+                // redirect
+                Session::flash('message', 'Successfully updated group');
+                return Redirect::to('groups');
+            }
         } else {
-            // store
-            $group = Group::find($id);
-            $group->name       = Input::get('name');
-            $group->save();
-
-            // redirect
-            Session::flash('message', 'Successfully updated group');
-            return Redirect::to('groups');
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
         }
+
     }
 
     /**
@@ -142,13 +171,17 @@ class GroupController extends BaseController {
      */
     public function destroy($id)
     {
-        $group = Group::find($id);
-        $group->delete();
+        if (Auth::check()) {
+            $group = Group::find($id);
+            $group->delete();
 
-        // redirect
-        Session::flash('message', 'Successfully deleted the group');
-        return Redirect::to('groups');
+            // redirect
+            Session::flash('message', 'Successfully deleted the group');
+            return Redirect::to('groups');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
-
-
 }
