+++ app/controllers/UserController.php
@@ -9,11 +9,17 @@ class UserController extends BaseController {
 
     public function index()
     {
-        $users = User::all();
-        return View::make('users.index')
-            ->with('users', $users)
-            ->with('title', 'Users')
-            ->with('count', count($users));
+        if (Auth::check()) {
+            $users = User::all();
+            return View::make('users.index')
+                ->with('users', $users)
+                ->with('title', 'Users')
+                ->with('count', count($users));
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -23,9 +29,15 @@ class UserController extends BaseController {
      */
     public function create()
     {
-        // load the create form (app/views/users/create.blade.php)
-        return View::make('users.create')
-                ->with('title', 'Users create');
+        if (Auth::check()) {
+            // load the create form (app/views/users/create.blade.php)
+            return View::make('users.create')
+                    ->with('title', 'Users create');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -35,37 +47,43 @@ class UserController extends BaseController {
      */
     public function store()
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'username'      => 'required',
-            'email'     => 'required',
-            'password'     => 'required',
-            'confirmation_code'     => 'required',
-            'remember_token'     => 'required',
-            'confirmed' => 'required',
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'username'      => 'required',
+                'email'     => 'required',
+                'password'     => 'required',
+                'confirmation_code'     => 'required',
+                'remember_token'     => 'required',
+                'confirmed' => 'required',
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('users/create')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('users/create')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $user = new User;
+                $user->username = Input::get('username');
+                $user->email = Input::get('email');
+                $user->password = Input::get('password');
+                $user->confirmation_code = Input::get('confirmation_code');
+                $user->remember_token = Input::get('remember_token');
+                $user->confirmed = Input::get('confirmed');
+                $user->save();
+
+                // redirect
+                Session::flash('message', 'Successfully created user');
+                return Redirect::to('users');
+            }
         } else {
-            // store
-            $user = new User;
-            $user->username = Input::get('username');
-            $user->email = Input::get('email');
-            $user->password = Input::get('password');
-            $user->confirmation_code = Input::get('confirmation_code');
-            $user->remember_token = Input::get('remember_token');
-            $user->confirmed = Input::get('confirmed');
-            $user->save();
-
-            // redirect
-            Session::flash('message', 'Successfully created user');
-            return Redirect::to('users');
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
         }
     }
 
@@ -77,13 +95,19 @@ class UserController extends BaseController {
      */
     public function show($id)
     {
-        // get the user
-        $user = User::find($id);
+        if (Auth::check()) {
+            // get the user
+            $user = User::find($id);
 
-        // show the view and pass the user to it
-        return View::make('users.show')
-            ->with('user', $user)
-            ->with('title', 'User show');
+            // show the view and pass the user to it
+            return View::make('users.show')
+                ->with('user', $user)
+                ->with('title', 'User show');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -94,13 +118,19 @@ class UserController extends BaseController {
      */
     public function edit($id)
     {
-        // get the user
-        $user = User::find($id);
+        if (Auth::check()) {
+            // get the user
+            $user = User::find($id);
 
-        // show the edit form and pass the user
-        return View::make('users.edit')
-            ->with('user', $user)
-            ->with('title', 'User edit');
+            // show the edit form and pass the user
+            return View::make('users.edit')
+                ->with('user', $user)
+                ->with('title', 'User edit');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -111,29 +141,35 @@ class UserController extends BaseController {
      */
     public function update($id)
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'name'       => 'required',
-            'email'      => 'required',
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'name'       => 'required',
+                'email'      => 'required',
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('users/' . $id . '/edit')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('users/' . $id . '/edit')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $user = User::find($id);
+                $user->name       = Input::get('name');
+                $user->email      = Input::get('email');
+                $user->save();
+
+                // redirect
+                Session::flash('message', 'Successfully updated user');
+                return Redirect::to('users');
+            }
         } else {
-            // store
-            $user = User::find($id);
-            $user->name       = Input::get('name');
-            $user->email      = Input::get('email');
-            $user->save();
-
-            // redirect
-            Session::flash('message', 'Successfully updated user');
-            return Redirect::to('users');
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
         }
     }
 
@@ -145,13 +181,17 @@ class UserController extends BaseController {
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        $user->delete();
+        if (Auth::check()) {
+            $user = User::find($id);
+            $user->delete();
 
-        // redirect
-        Session::flash('message', 'Successfully deleted the user');
-        return Redirect::to('users');
+            // redirect
+            Session::flash('message', 'Successfully deleted the user');
+            return Redirect::to('users');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
-
-
 }
