+++ app/controllers/WorkController.php
@@ -15,12 +15,18 @@ class WorkController extends BaseController {
      */
     public function index()
     {
-        $works = Work::all();
+        if (Auth::check()) {
+            $works = Work::all();
 
-        return View::make('works.index')
-            ->with('works', $works)
-            ->with('title', 'Works')
-            ->with('count', count($works));
+            return View::make('works.index')
+                ->with('works', $works)
+                ->with('title', 'Works')
+                ->with('count', count($works));
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -30,9 +36,15 @@ class WorkController extends BaseController {
      */
     public function create()
     {
-        // load the create form (app/views/works/create.blade.php)
-        return View::make('works.create')
-            ->with('title', 'Work create');
+        if (Auth::check()) {
+            // load the create form (app/views/works/create.blade.php)
+            return View::make('works.create')
+                ->with('title', 'Work create');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -42,21 +54,22 @@ class WorkController extends BaseController {
      */
     public function store()
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'title'      => 'required',
-            'reference'  => 'required',
-            'media'      => 'required',
-            'dimensions' => 'required',
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'title'      => 'required',
+                'reference'  => 'required',
+                'media'      => 'required',
+                'dimensions' => 'required',
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('works/create')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('works/create')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
         } else {
             // store
             $work = new Work;
@@ -71,6 +84,11 @@ class WorkController extends BaseController {
             Session::flash('message', 'Successfully created work');
             return Redirect::to('works');
         }
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -81,13 +99,19 @@ class WorkController extends BaseController {
      */
     public function show($id)
     {
-        // get the work
-        $work = Work::find($id);
+        if (Auth::check()) {
+            // get the work
+            $work = Work::find($id);
 
-        // show the view and pass the nerd to it
-        return View::make('works.show')
-            ->with('work', $work)
-            ->with('title', 'Work show');
+            // show the view and pass the nerd to it
+            return View::make('works.show')
+                ->with('work', $work)
+                ->with('title', 'Work show');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -98,13 +122,19 @@ class WorkController extends BaseController {
      */
     public function edit($id)
     {
-        // get the work
-        $work = Work::find($id);
+        if (Auth::check()) {
+            // get the work
+            $work = Work::find($id);
 
-        // show the edit form and pass the work
-        return View::make('works.edit')
-            ->with('work', $work)
-            ->with('title', 'Work edit');
+            // show the edit form and pass the work
+            return View::make('works.edit')
+                ->with('work', $work)
+                ->with('title', 'Work edit');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
 
     /**
@@ -115,34 +145,40 @@ class WorkController extends BaseController {
      */
     public function update($id)
     {
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $rules = array(
-            'title'       => 'required',
-            'reference'      => 'required',
-            'media'       => 'required',
-            'dimensions'      => 'required',
-            'work_date' => 'required'
-        );
-        $validator = Validator::make(Input::all(), $rules);
+        if (Auth::check()) {
+            // validate
+            // read more on validation at http://laravel.com/docs/validation
+            $rules = array(
+                'title'       => 'required',
+                'reference'      => 'required',
+                'media'       => 'required',
+                'dimensions'      => 'required',
+                'work_date' => 'required'
+            );
+            $validator = Validator::make(Input::all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('works/' . $id . '/edit')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('works/' . $id . '/edit')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $work = Work::find($id);
+                $work->title       = Input::get('title');
+                $work->reference       = Input::get('reference');
+                $work->media      = Input::get('media');
+                $work->work_date = Input::get('work_date');
+                $work->save();
+
+                // redirect
+                Session::flash('message', 'Successfully updated work');
+                return Redirect::to('works');
+            }
         } else {
-            // store
-            $work = Work::find($id);
-            $work->title       = Input::get('title');
-            $work->reference       = Input::get('reference');
-            $work->media      = Input::get('media');
-            $work->work_date = Input::get('work_date');
-            $work->save();
-
-            // redirect
-            Session::flash('message', 'Successfully updated work');
-            return Redirect::to('works');
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
         }
     }
 
@@ -154,12 +190,18 @@ class WorkController extends BaseController {
      */
     public function destroy($id)
     {
-        $work = Work::find($id);
-        $work->delete();
+        if (Auth::check()) {
+            $work = Work::find($id);
+            $work->delete();
 
-        // redirect
-        Session::flash('message', 'Successfully deleted the work');
-        return Redirect::to('works');
+            // redirect
+            Session::flash('message', 'Successfully deleted the work');
+            return Redirect::to('works');
+        } else {
+            // User is not logged in
+            Session::flash('message', 'Please log in');
+            return Redirect::to('/');
+        }
     }
